Updates
Added new images for background and hearts.

App.js modified the overlay background and put the keyup event inside the Start game listener. (This was to prevent errors when reloading with ctrl + r)

Game.js changed phrases to reflect new theme. Added logic to change background depending on a win or loss.